+++ docs/changelog.md
@@ -1,6 +1,13 @@
 # TBD
+
+# 0.2.3
+### Features
+* Added the ability to specify arbitrary numbers of participants with EL/CL combos, and default to one Geth+Nimbus participant
+* Added instructions to the README for configuring the module
+
 ### Fixes
 * Get rid of the 300-second delay in the generated CL genesis
+* Added the Lighthouse validator node
 
 # 0.2.2
 ### Features